fix: 상태이상 관련 수정 작업

+++ src/main/kotlin/org/beobma/classWarPlugin/manager/PlayerManager.kt
@@ -11,9 +11,7 @@ import org.beobma.classWarPlugin.manager.GameClassManager.toItemStack
 import org.beobma.classWarPlugin.manager.UtilManager.sendMiniMessage
 import org.beobma.classWarPlugin.util.DamageCalculator
 import org.beobma.classWarPlugin.util.DamageType
-import org.beobma.classWarPlugin.util.DamageType.Normal
 import org.beobma.classWarPlugin.util.DamageType.StatusAbnormality
-import org.beobma.classWarPlugin.util.DamageType.True
 import org.bukkit.Bukkit
 import org.bukkit.Material
 import org.bukkit.entity.Player
@@ -86,7 +84,12 @@ object PlayerManager {
         }
     }
 
-    fun PlayerData.damage(damage: Double, damageType: DamageType, damager: PlayerData, isInvincibilityTimeIgnore: Boolean = true) {
+    fun PlayerData.damage(
+        damage: Double,
+        damageType: DamageType,
+        damager: PlayerData,
+        isInvincibilityTimeIgnore: Boolean = true
+    ) {
         if (damage <= 0.0) {
             return
         }
@@ -146,7 +149,7 @@ object PlayerManager {
         if (finalDamage < 0) {
             return
         }
-        player.heal(finalDamage.toDouble())
+        player.heal(finalDamage)
     }
 
     fun EntityData.damage(
